feat: improve websockets server and add wip d3 frontend
Add more capabilites to the websocket server and to the simulation.
The simulation can now have a different number of philosophers.
The simulation now can be started and stopped with websocket messages.
Add wip d3 frontend page for the simulation. (still needs quite a bit of
work, currently there are only some static shapes drawn on the screen.
The simulation is not running on cli anymore, we could create a new
cli.js that will run it on the command line if needed.